Se posicionan los personajes de manera independiente a PesonajeDef, sino que a travez de personajes pertenecientes a cada jugador respectivo

+++ trunk/src/modelo/Vector2f.h
@@ -26,7 +26,7 @@ public:
 	Vector2f& operator-=(const Vector2f& v2);
 	Vector2f operator*(float escalar);
 	Vector2f& operator*=(float escalar);
-    void setCoordenada(int x, int y);
+    void setCoordenada(double x, double y);
     friend ostream& operator<<(ostream &o, const Vector2f &c);
     string toString();
 };
